+++ packages/core/src/themes/context.tsx
@@ -5,8 +5,16 @@ import { mergeTheme } from './utils'
 
 export const ThemeContext = createContext(defaultTheme as CompleteThemeSpec)
 
-export const ThemeProvider = ({ theme, children }: { theme?: ThemeSpec; children: ReactNode }) => {
-    const mergedTheme = useMemo(() => mergeTheme(defaultTheme, theme) as CompleteThemeSpec, [theme])
+export const ThemeProvider = ({
+    baseTheme,
+    theme,
+    children,
+}: {
+    baseTheme: CompleteThemeSpec
+    theme?: ThemeSpec
+    children: ReactNode
+}) => {
+    const mergedTheme = useMemo(() => mergeTheme(baseTheme, theme) as CompleteThemeSpec, [theme])
     return <ThemeContext.Provider value={mergedTheme}>{children}</ThemeContext.Provider>
 }