+++ packages/core/src/themes/utils.ts
@@ -8,23 +8,29 @@ import { useMemo } from 'react'
 // turn an array of string into a single space-separate className string
 export const composeClassName = function (names: Array<string | undefined>) {
     const result = names
-        .filter(x => x !== undefined)
         .map(x => camelCase(x))
+        .filter(x => x !== undefined)
         .join(' ')
     return result === '' ? undefined : result
 }
 
 export const addColor = (style: CssProps | undefined, color: string) => {
     if (!style) return { fill: color, stroke: color }
-    let result = cloneDeep(style)
-    if (!result.fill) result = merge(result, { fill: color })
-    if (!result.stroke) result = merge(result, { stroke: color })
+    const result = cloneDeep(style)
+    if (!result.fill) {
+        result.fill = color
+    }
+    if (!result.stroke) {
+        result.stroke = color
+    }
     return result
 }
 
 export const addOpacity = (style: CssProps | undefined, opacity: number) => {
     if (!style) return { opacity }
-    return merge(cloneDeep(style), { opacity })
+    const result = cloneDeep(style)
+    result.opacity = opacity
+    return result
 }
 
 export const mergeTheme = (baseTheme: ThemeSpec, customTheme?: ThemeSpec) => {