+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+-   [packages/core]: prop `baseTheme` for `Chart` to enable resetting the entire
+    theme
 -   [packages/themes]: theme pieces for text, lines, and ticks
 -   [packages/annotation]: symbols for horizontal and vertical golden rectangles
 -   [packages/core]: object `containerRef` in dimensions context